Refactor relay_protocol.c into modular files
